refactor: 토큰 관련 비즈니스 로직, filter 로직 구분 (#11)
